feat : 게시글 조회 (무한스크롤) 기능 구현 및 Restdocs (#197)
* feat : #194 게시글 무한스크롤 조회 기능 구현

* docs : #194 PostControllerTest, restdocs 작성

* docs : #194 PostControllerTest, restdocs 작성

* feat : #194 게시글 소유권 로직 추가

+++ backend/src/main/java/com/daily/daily/post/service/PostService.java
@@ -1,17 +1,21 @@
 package com.daily.daily.post.service;
 
+import com.daily.daily.common.exception.UnauthorizedAccessException;
 import com.daily.daily.common.service.S3StorageService;
 import com.daily.daily.member.domain.Member;
 import com.daily.daily.member.exception.MemberNotFoundException;
 import com.daily.daily.member.repository.MemberRepository;
 import com.daily.daily.post.domain.Post;
 import com.daily.daily.post.dto.PostReadResponseDTO;
+import com.daily.daily.post.dto.PostReadSliceResponseDTO;
 import com.daily.daily.post.dto.PostWriteRequestDTO;
 import com.daily.daily.post.dto.PostWriteResponseDTO;
 import com.daily.daily.post.exception.PostNotFoundException;
 import com.daily.daily.post.repository.PostLikeRepository;
 import com.daily.daily.post.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -51,10 +55,11 @@ public class PostService {
         return PostWriteResponseDTO.from(savedPost);
     }
 
-    public PostWriteResponseDTO update(Long postId, PostWriteRequestDTO postWriteRequestDTO, MultipartFile pageImage) {
+    public PostWriteResponseDTO update(Long memberId, Long postId, PostWriteRequestDTO postWriteRequestDTO, MultipartFile pageImage) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(PostNotFoundException::new);
 
+        validateAuthorityToPost(post, memberId);
         post.updateContent(postWriteRequestDTO.getContent());
         hashtagService.updateHashtagsInPost(post, postWriteRequestDTO.getHashtags());
 
@@ -78,7 +83,33 @@ public class PostService {
         return PostReadResponseDTO.from(post, likeCount);
     }
 
-    public void delete(Long postId) {
+    public PostReadSliceResponseDTO findSlice(Pageable pageable) {
+        Slice<Post> posts = postRepository.find(pageable);
+
+        /**
+         * TODO: 이렇게 하면 좋아요 갯수를 조회하기 위해 N+1쿼리가 발생한다. 따라서 최적화가 필요한데... 어떻게 처리할 지 고민이다.
+         */
+        for (Post post : posts) {
+            Long likeCount = likeRepository.countByPost(post);
+            post.setLikeCount(likeCount);
+        }
+
+        return PostReadSliceResponseDTO.from(posts);
+    }
+
+    public void delete(Long memberId, Long postId) {
+        Post post = postRepository.findById(postId)
+                .orElseThrow(PostNotFoundException::new);
+
+        validateAuthorityToPost(post, memberId);
+
         postRepository.deleteById(postId);
+    }
+
+
+    private void validateAuthorityToPost(Post post, Long writerId) {
+        if (!post.isWrittenBy(writerId)) {
+            throw new UnauthorizedAccessException();
+        }
     }
 }
